consolidate command to uniques
